% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{show,band-method}
\alias{show,band-method}
\title{Show Method for Survival Band}
\usage{
\S4method{show}{band}(object)
}
\arguments{
\item{object}{An object of class \code{band}.}
}
\description{
Show Method for Survival Band
}
