% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSample.R
\name{Bands.OneSample}
\alias{Bands.OneSample}
\title{Estimate 1-Sample Confidence Bands}
\usage{
Bands.OneSample(
  time,
  status,
  alpha = 0.05,
  paths = 1000,
  method = "EP",
  lower = NULL,
  upper = NULL
)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, 1 for an event, 0 for censoring.}

\item{alpha}{Alpha level.}

\item{paths}{Sample paths.}

\item{method}{Either equiprecision "EP" or Hall-Wellner "HW".}

\item{lower}{Lower endpoint of interval over which confidence bands are sought.}

\item{upper}{Upper endpoint of interval over which confidence bands are sought.}
}
\value{
Object of class band, containing a table with the lower and upper
  confidence bands.
}
\description{
Constructs the confidence band on the log cumulative hazard (CH) scale,
then transforms onto the survival scale. The critical value, on the
log CH scale, is determined via simulation.
}
