% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bands.R
\name{draw.bands}
\alias{draw.bands}
\title{Simple interactive chromosome maps}
\usage{
draw.bands(df, bands = NULL, labels = FALSE)
}
\arguments{
\item{df}{A data frame of chromosome segments}

\item{bands}{Specificy which chomosomes to show and in what order}

\item{labels}{Set true to show segment labels}
}
\value{
A plotly object
}
\description{
Simple interactive chromosome maps
}
\examples{

samples <- 250
bands <- paste('Chromosome', 0:9)
starts <- sample(1:500000, samples, replace=TRUE)
ends <- sapply(starts, function(x) x+sample(1:10000, 1))
df <- data.frame(band = sample(bands, samples, replace=TRUE),
                 start = starts,
                 end = ends,
                 stringsAsFactors=FALSE)

draw.bands(df)

}
