% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSample.R
\name{oneSampleBand}
\alias{oneSampleBand}
\title{Estimate 1-Sample Confidence Bands}
\usage{
oneSampleBand(time, status, sig = 0.05, paths = 1000, weights = "ep",
  L = NULL, U = NULL)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, 1 for an event, 0 for censoring.}

\item{sig}{Significance level.}

\item{paths}{Sample paths.}

\item{weights}{Either equiprecision "ep" or Hall-Wellner "hw".}

\item{L}{Lower endpoint of interval over which confidence bands are sought.}

\item{U}{Upper endpoint of interval over which confidence bands are sought.}
}
\value{
Object of class band, containing a table with the lower and upper
  confidence bands.
}
\description{
Constructs the confidence band on the log cumulative hazard (CH) scale,
then transforms onto the survival scale. The critical value, on the
log CH scale, is determined via simulation.
}
