% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{plot.band}
\alias{plot.band}
\title{Plot Method for Survival Band}
\usage{
\method{plot}{band}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{band}.}

\item{y}{Unused.}

\item{...}{Unused.}
}
\description{
Plot method for objects of class \code{band}.
}
