% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.OneSample}
\alias{Plot.OneSample}
\title{Plot 1-Sample Confidence Band}
\usage{
Plot.OneSample(band)
}
\arguments{
\item{band}{Object of class band.}
}
\value{
An object of class `ggplot`.
}
\description{
Plot 1-Sample Confidence Band
}
