% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{Plot.TwoSample}
\alias{Plot.TwoSample}
\title{Plot 2-Sample Confidence Band}
\usage{
Plot.TwoSample(band)
}
\arguments{
\item{band}{Object of class band.}
}
\value{
An object of class `ggplot`.
}
\description{
Plot 2-Sample Confidence Band
}
