% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{PlotBands}
\alias{PlotBands}
\title{Master Plotting Function}
\usage{
PlotBands(band)
}
\arguments{
\item{band}{Object of class band.}
}
\value{
An object of class `ggplot`.
}
\description{
Master Plotting Function
}
