% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.band}
\alias{print.band}
\title{Print Method for Survival Band}
\usage{
\method{print}{band}(x, ...)
}
\arguments{
\item{x}{An object of class \code{band}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{band}.
}
