% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSample.R
\name{SamplePath}
\alias{SamplePath}
\title{Sample Path of Survival Process}
\usage{
SamplePath(times, events, nar, surv, weight = NULL)
}
\arguments{
\item{times}{Distinct observed event times.}

\item{events}{Number of events at each time.}

\item{nar}{Number at risk.}

\item{surv}{Survival probability.}

\item{weight}{Weight to give each time point.}
}
\value{
Numeric vector of the same length as `times`.
}
\description{
Perturbation realization of a sample path from \eqn{\sqrt{n}w(t)\{\hat{S}(t)-S(t)\}}.
}
