% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KS.R
\name{survKS}
\alias{survKS}
\title{2-Sample Kolmogorov Smirnov Test}
\usage{
survKS(time, status, arm, paths = 1000, weights = "ep")
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, 1 for an event, 0 for censoring.}

\item{arm}{Treatment indicator, 1 for target group, 0 for reference.}

\item{paths}{Sample paths.}

\item{weights}{Either equiprecision "ep" or Hall-Wellner "hw".}
}
\value{
List containing the KS statistic and the p-value.
}
\description{
2-Sample Kolmogorov Smirnov Test
}
