% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSample.R
\name{twoSampleBand}
\alias{twoSampleBand}
\title{Estimate 2-Sample Confidence Bands}
\usage{
twoSampleBand(time, status, arm, sig = 0.05, paths = 1000,
  weights = "ep", L = NULL, U = NULL)
}
\arguments{
\item{time}{Observation time.}

\item{status}{Status, 1 for an event, 0 for censoring.}

\item{arm}{Treatment indicator, 1 for target group, 0 for reference.}

\item{sig}{Significance level.}

\item{paths}{Sample paths.}

\item{weights}{Either equiprecision "ep" or Hall-Wellner "hw".}

\item{L}{Lower endpoint of interval over which confidence bands are sought.}

\item{U}{Upper endpoint of interval over which confidence bands are sought.}
}
\value{
Object of class band, containing the lower and upper confidence bands.
}
\description{
Estimate 2-Sample Confidence Bands
}
