% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{band-class}
\alias{band-class}
\title{Fitted Survival Bands}
\description{
Defines the object class returned by fitting functions.
}
\section{Slots}{

\describe{
\item{\code{Alpha}}{Alpha level.}

\item{\code{Crit}}{Critical value.}

\item{\code{Paths}}{Number of simulated sample paths.}

\item{\code{Pvalue}}{P-value for the difference of survival curves.}

\item{\code{Samples}}{Number of samples.}

\item{\code{Table}}{Tabulated survival function with confidence bands.}
}}

